backend/requirements.txt
flask==3.0.3                # A lightweight WSGI web application framework in Python.
flask-cors==5.0.0           # A Flask extension for handling Cross-Origin Resource Sharing (CORS).
basicsr==1.4.2              # BasicSR is a PyTorch-based image and video super-resolution toolbox.
numpy==1.26.3               # A fundamental package for numerical computation in Python.
opencv-python==4.10.0.84    # OpenCV library for computer vision tasks in Python.
opencv-python-headless==4.10.0.84 # OpenCV library for computer vision tasks without GUI support.
Pillow==9.5.0               # A Python Imaging Library (PIL) fork that adds image processing capabilities.
pyiqa==0.1.13               # A library for perceptual image quality assessment using deep learning.
realesrgan==0.3.0           # A Real-World Super Resolution Generative Adversarial Network for image enhancement.
torch==2.4.1+cu118          # PyTorch, an open-source machine learning framework for tensor computation.
torchaudio==2.4.1+cu118     # PyTorch's audio processing library, providing functionalities for audio processing.
torchvision==0.19.1         # PyTorch's computer vision library, providing datasets, model architectures, and image transformations.
